\name{params.dilution}
\alias{params.dilution}
\docType{data}
\title{ Parameters for the Dilution dataset }
\description{
  PDNN parameters for the Dilution dataset
}
\usage{data(params.dilution)}
\format{
  The format is:
List of 6
 $ lambda      :List of 12625 (probesets) ...
 $ Bs          : num [1:4] 111.9  57.3 120.5  50.1
 $ Ns          : num [1:4] 2967 2998 2992 2999
 $ Fs          : num [1:4] 0.607 0.662 0.600 0.656
 $ names.abatch: chr [1:12625] "1000\\_at" "1001\\_at" "1002\\_f\\_at" ...
 $ names.i     : int [1:12625] 1 2 3 4 5 6 7 8 9 10 ...
}
\details{
  These data are provided to reduce the time needed to run the vignette.
}
\examples{
data(params.dilution)
}
\keyword{datasets}
